ref(crons): Split processing_errors into 2 modules
Splits out the errors from the manager. Removes the
CheckinProcessErrorsManager class (which had no state) and updated the
function names to be more clear
