+++ src/sentry/monitors/endpoints/project_monitor_processing_errors_index.py
@@ -9,10 +9,8 @@ from sentry.api.serializers import serialize
 from sentry.apidocs.constants import RESPONSE_FORBIDDEN, RESPONSE_NOT_FOUND, RESPONSE_UNAUTHORIZED
 from sentry.apidocs.parameters import GlobalParams, MonitorParams
 from sentry.apidocs.utils import inline_sentry_response_serializer
-from sentry.monitors.processing_errors import (
-    CheckinProcessErrorsManager,
-    CheckinProcessingErrorData,
-)
+from sentry.monitors.processing_errors.errors import CheckinProcessingErrorData
+from sentry.monitors.processing_errors.manager import get_errors_for_monitor
 from sentry.utils.auth import AuthenticatedHttpRequest
 
 from .base import ProjectMonitorEndpoint
@@ -46,9 +44,7 @@ class ProjectMonitorProcessingErrorsIndexEndpoint(ProjectMonitorEndpoint):
         """
         Retrieves checkin processing errors for a monitor
         """
-        paginator = SequencePaginator(
-            list(enumerate(CheckinProcessErrorsManager().get_for_monitor(monitor)))
-        )
+        paginator = SequencePaginator(list(enumerate(get_errors_for_monitor(monitor))))
 
         return self.paginate(
             request=request,
