+++ src/sentry/monitors/serializers.py
@@ -17,7 +17,10 @@ from sentry.monitors.models import (
     MonitorIncident,
     MonitorStatus,
 )
-from sentry.monitors.processing_errors import CheckinProcessingError, CheckinProcessingErrorData
+from sentry.monitors.processing_errors.errors import (
+    CheckinProcessingError,
+    CheckinProcessingErrorData,
+)
 from sentry.monitors.utils import fetch_associated_groups
 from sentry.monitors.validators import IntervalNames
 from sentry.types.actor import Actor
