+++ tests/sentry/monitors/processing_errors/test_manager.py
@@ -1,69 +1,62 @@
 from unittest import mock
 
-from sentry.monitors.processing_errors import (
-    CheckinProcessErrorsManager,
+from sentry.monitors.processing_errors.errors import (
     CheckinProcessingError,
     CheckinValidationError,
     ProcessingError,
     ProcessingErrorType,
+)
+from sentry.monitors.processing_errors.manager import (
+    _get_cluster,
+    build_error_identifier,
+    delete_error,
+    get_errors_for_monitor,
+    get_errors_for_projects,
     handle_processing_errors,
+    store_error,
 )
 from sentry.monitors.testutils import build_checkin_item, build_checkin_processing_error
 from sentry.testutils.cases import TestCase
 
 
-class ProcessingErrorTest(TestCase):
-    def test(self):
-        error = ProcessingError(ProcessingErrorType.CHECKIN_INVALID_GUID, {"some": "data"})
-        recreated_error = ProcessingError.from_dict(error.to_dict())
-        assert recreated_error.type == error.type
-        assert recreated_error.data == error.data
-
-
-class CheckinProcessingErrorTest(TestCase):
-    def test(self):
-        item = build_checkin_item()
-        error = CheckinProcessingError(
-            [ProcessingError(ProcessingErrorType.MONITOR_DISABLED, {"some": "data"})],
-            item,
-        )
-        recreated_error = CheckinProcessingError.from_dict(error.to_dict())
-        assert error == recreated_error
+def assert_processing_errors_equal(
+    error_1: CheckinProcessingError,
+    error_2: CheckinProcessingError,
+):
+    assert error_1.errors == error_2.errors
+    assert error_2.checkin == error_2.checkin
 
 
 class CheckinProcessErrorsManagerTest(TestCase):
     def test_store_with_monitor(self):
         monitor = self.create_monitor()
-        manager = CheckinProcessErrorsManager()
         processing_error = build_checkin_processing_error()
-        manager.store(processing_error, monitor)
-        fetched_processing_error = manager.get_for_monitor(monitor)
+        store_error(processing_error, monitor)
+        fetched_processing_error = get_errors_for_monitor(monitor)
         assert len(fetched_processing_error) == 1
-        self.assert_processing_errors_equal(processing_error, fetched_processing_error[0])
+        assert_processing_errors_equal(processing_error, fetched_processing_error[0])
 
     def test_store_with_slug_exists(self):
         monitor = self.create_monitor()
-        manager = CheckinProcessErrorsManager()
         processing_error = build_checkin_processing_error(
             message_overrides={"project_id": self.project.id},
             payload_overrides={"monitor_slug": monitor.slug},
         )
-        manager.store(processing_error, None)
-        fetched_processing_error = manager.get_for_monitor(monitor)
+        store_error(processing_error, None)
+        fetched_processing_error = get_errors_for_monitor(monitor)
         assert len(fetched_processing_error) == 1
-        self.assert_processing_errors_equal(processing_error, fetched_processing_error[0])
+        assert_processing_errors_equal(processing_error, fetched_processing_error[0])
 
     def test_store_with_slug_not_exist(self):
-        manager = CheckinProcessErrorsManager()
         processing_error = build_checkin_processing_error(
             message_overrides={"project_id": self.project.id},
             payload_overrides={"monitor_slug": "hi"},
         )
 
-        manager.store(processing_error, None)
-        fetched_processing_error = manager.get_for_projects([self.project])
+        store_error(processing_error, None)
+        fetched_processing_error = get_errors_for_projects([self.project])
         assert len(fetched_processing_error) == 1
-        self.assert_processing_errors_equal(processing_error, fetched_processing_error[0])
+        assert_processing_errors_equal(processing_error, fetched_processing_error[0])
 
     def test_store_max(self):
         monitor = self.create_monitor()
@@ -84,36 +77,26 @@ class CheckinProcessErrorsManagerTest(TestCase):
                 payload_overrides={"monitor_slug": monitor.slug},
             ),
         ]
-        manager = CheckinProcessErrorsManager()
-        with mock.patch("sentry.monitors.processing_errors.MAX_ERRORS_PER_SET", new=2):
+        with mock.patch("sentry.monitors.processing_errors.manager.MAX_ERRORS_PER_SET", new=2):
             for error in processing_errors:
-                manager.store(error, monitor)
+                store_error(error, monitor)
 
-        retrieved_errors = manager.get_for_monitor(monitor)
+        retrieved_errors = get_errors_for_monitor(monitor)
         assert len(retrieved_errors) == 2
-        self.assert_processing_errors_equal(processing_errors[2], retrieved_errors[0])
-        self.assert_processing_errors_equal(processing_errors[1], retrieved_errors[1])
-
-    def assert_processing_errors_equal(
-        self, error_1: CheckinProcessingError, error_2: CheckinProcessingError
-    ):
-        assert error_1.errors == error_2.errors
-        assert error_2.checkin == error_2.checkin
+        assert_processing_errors_equal(processing_errors[2], retrieved_errors[0])
+        assert_processing_errors_equal(processing_errors[1], retrieved_errors[1])
 
     def test_get_for_monitor_empty(self):
-        manager = CheckinProcessErrorsManager()
         monitor = self.create_monitor()
-        assert len(manager.get_for_monitor(monitor)) == 0
+        assert len(get_errors_for_monitor(monitor)) == 0
 
     def test_get_for_project(self):
-        manager = CheckinProcessErrorsManager()
-        assert len(manager.get_for_projects([self.project])) == 0
+        assert len(get_errors_for_projects([self.project])) == 0
 
     def test_get_missing_data(self):
         # Validate that we don't error if a processing error has expired but is still
         # in the set
         monitor = self.create_monitor()
-        manager = CheckinProcessErrorsManager()
         processing_errors = [
             build_checkin_processing_error(
                 [ProcessingError(ProcessingErrorType.CHECKIN_INVALID_GUID, {"guid": "bad"})],
@@ -127,34 +110,32 @@ class CheckinProcessErrorsManagerTest(TestCase):
             ),
         ]
         for processing_error in processing_errors:
-            manager.store(processing_error, monitor)
-        redis = manager._get_cluster()
-        redis.delete(manager.build_error_identifier(processing_errors[0].id))
-        fetched_processing_error = manager.get_for_monitor(monitor)
+            store_error(processing_error, monitor)
+        redis = _get_cluster()
+        redis.delete(build_error_identifier(processing_errors[0].id))
+        fetched_processing_error = get_errors_for_monitor(monitor)
         assert len(fetched_processing_error) == 1
-        self.assert_processing_errors_equal(processing_errors[1], fetched_processing_error[0])
+        assert_processing_errors_equal(processing_errors[1], fetched_processing_error[0])
 
     def test_delete_for_monitor(self):
-        manager = CheckinProcessErrorsManager()
         monitor = self.create_monitor()
         processing_error = build_checkin_processing_error(
             message_overrides={"project_id": self.project.id},
             payload_overrides={"monitor_slug": monitor.slug},
         )
-        manager.store(processing_error, monitor)
-        assert len(manager.get_for_monitor(monitor)) == 1
-        manager.delete(self.project, processing_error.id)
-        assert len(manager.get_for_monitor(monitor)) == 0
+        store_error(processing_error, monitor)
+        assert len(get_errors_for_monitor(monitor)) == 1
+        delete_error(self.project, processing_error.id)
+        assert len(get_errors_for_monitor(monitor)) == 0
 
     def test_delete_for_project(self):
-        manager = CheckinProcessErrorsManager()
         processing_error = build_checkin_processing_error(
             message_overrides={"project_id": self.project.id},
         )
-        manager.store(processing_error, None)
-        assert len(manager.get_for_projects([self.project])) == 1
-        manager.delete(self.project, processing_error.id)
-        assert len(manager.get_for_projects([self.project])) == 0
+        store_error(processing_error, None)
+        assert len(get_errors_for_projects([self.project])) == 1
+        delete_error(self.project, processing_error.id)
+        assert len(get_errors_for_projects([self.project])) == 0
 
 
 class HandleProcessingErrorsTest(TestCase):
@@ -170,8 +151,7 @@ class HandleProcessingErrorsTest(TestCase):
             ),
             exception,
         )
-        manager = CheckinProcessErrorsManager()
-        errors = manager.get_for_monitor(monitor)
+        errors = get_errors_for_monitor(monitor)
         assert not errors
         with self.feature("organizations:crons-write-user-feedback"):
             handle_processing_errors(
@@ -180,5 +160,5 @@ class HandleProcessingErrorsTest(TestCase):
                 ),
                 exception,
             )
-        errors = manager.get_for_monitor(monitor)
+        errors = get_errors_for_monitor(monitor)
         assert len(errors) == 1
